[DEV] Improved file organisation by grouping files into packages.
