+++ src/main/java/data/database/Database.java
@@ -8,6 +8,13 @@ import models.Account;
 import models.Code;
 import models.User;
 import utilities.Utilities;
+import utilities.sqliteutilities.SQLiteHelper;
+import utilities.sqliteutilities.argumentsetters.IntegerSetter;
+import utilities.sqliteutilities.argumentsetters.StringSetter;
+import utilities.sqliteutilities.retrievers.CodeRetriever;
+import utilities.sqliteutilities.retrievers.IntegerRetriever;
+import utilities.sqliteutilities.retrievers.ListRetriever;
+import utilities.sqliteutilities.retrievers.StringRetriever;
 
 import javax.crypto.SecretKey;
 import javax.crypto.SecretKeyFactory;
