+++ src/main/java/utilities/sqliteutilities/argumentsetters/ArgumentSetter.java
@@ -1,4 +1,4 @@
-package data;
+package utilities.sqliteutilities.argumentsetters;
 
 import java.sql.PreparedStatement;
