+++ src/main/java/utilities/sqliteutilities/retrievers/CodeRetriever.java
@@ -1,7 +1,8 @@
-package data;
+package utilities.sqliteutilities.retrievers;
 
 import javafx.util.Callback;
 import models.Code;
+import utilities.sqliteutilities.retrievers.Retriever;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
