+++ src/main/java/utilities/sqliteutilities/retrievers/ListRetriever.java
@@ -1,6 +1,7 @@
-package data;
+package utilities.sqliteutilities.retrievers;
 
 import javafx.util.Callback;
+import utilities.sqliteutilities.retrievers.Retriever;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
