+++ src/main/java/utilities/sqliteutilities/retrievers/StringRetriever.java
@@ -1,4 +1,4 @@
-package data;
+package utilities.sqliteutilities.retrievers;
 
 import javafx.util.Callback;
